+++ src/Controller/Admin/BackupsController.php
@@ -100,7 +100,7 @@ class BackupsController extends AppController
             if ($backup->execute($this->getRequest()->getData())) {
                 $this->Flash->success(I18N_OPERATION_OK);
 
-                return $this->redirect(['action' => 'index']);
+                return $this->redirect($this->referer(['action' => 'index']));
             }
 
             $this->Flash->error(I18N_OPERATION_NOT_OK);
@@ -123,7 +123,7 @@ class BackupsController extends AppController
         call_user_func([$this->BackupManager, $filename ? 'delete' : 'deleteAll'], $filename);
         $this->Flash->success(I18N_OPERATION_OK);
 
-        return $this->redirect(['action' => 'index']);
+        return $this->redirect($this->referer(['action' => 'index']));
     }
 
     /**
@@ -173,7 +173,7 @@ class BackupsController extends AppController
 
         $this->Flash->success(I18N_OPERATION_OK);
 
-        return $this->redirect(['action' => 'index']);
+        return $this->redirect($this->referer(['action' => 'index']));
     }
 
     /**
@@ -189,6 +189,6 @@ class BackupsController extends AppController
         $this->BackupManager->send($this->getFilename($filename), getConfigOrFail('email.webmaster'));
         $this->Flash->success(I18N_OPERATION_OK);
 
-        return $this->redirect(['action' => 'index']);
+        return $this->redirect($this->referer(['action' => 'index']));
     }
 }