+++ src/Controller/Admin/BannersController.php
@@ -155,7 +155,7 @@ class BannersController extends AppController
             if ($this->Banners->save($banner)) {
                 $this->Flash->success(I18N_OPERATION_OK);
 
-                return $this->redirect(['action' => 'index', $banner->get('position_id')]);
+                return $this->redirect($this->referer(['action' => 'index']));
             }
 
             $this->Flash->error(I18N_OPERATION_NOT_OK);
@@ -182,10 +182,9 @@ class BannersController extends AppController
     public function delete($id)
     {
         $this->getRequest()->allowMethod(['post', 'delete']);
-        $banner = $this->Banners->get($id);
-        $this->Banners->deleteOrFail($banner);
+        $this->Banners->deleteOrFail($this->Banners->get($id));
         $this->Flash->success(I18N_OPERATION_OK);
 
-        return $this->redirect(['action' => 'index', $banner->get('position_id')]);
+        return $this->redirect($this->referer(['action' => 'index']));
     }
 }