+++ src/Controller/Admin/LogsController.php
@@ -142,6 +142,6 @@ class LogsController extends AppController
         }
         call_user_func([$this->Flash, $method], $message);
 
-        return $this->redirect(['action' => 'index']);
+        return $this->redirect($this->referer(['action' => 'index']));
     }
 }